renamed locations.js and updated routes. Started schema creation process.
