+++ app_server/models/service_providers.js
@@ -1,0 +1,2 @@
+var mongoose = require( 'mongoose' );
+var sprovidersSchema = new mongoose.Schema({ });
