+++ app_server/routes/index.js
@@ -1,21 +1,21 @@
 var express = require('express');
 var router = express.Router();
 // var ctrlMain = require('../controllers/main'); //3.4
-var ctrlLocations = require('../controllers/locations');
+var ctrlServices = require('../controllers/services');
 // var ctrlOthers = require('../controllers/others');
 
 /* Locations pages */
-router.get('/', ctrlLocations.login);
-router.get('/main', ctrlLocations.main);
+router.get('/', ctrlServices.login);
+router.get('/main', ctrlServices.main);
 
 //get pets
-router.get('/pets', ctrlLocations.pets); 
+router.get('/pets', ctrlServices.pets); 
 
 //get rides
-router.get('/rides', ctrlLocations.rides);
+router.get('/rides', ctrlServices.rides);
 
 //get goods
-router.get('/goods', ctrlLocations.goods);
+router.get('/goods', ctrlServices.goods);
 /* GET home page. */
 // var homepageController = function (req, res) {
 // 	res.render('index', {title: 'Express'});
